Rename internal package name to avoid triggering vuln scan.
