+++ scripts/functions-discover-tests/fixtures/yarn-workspaces/packages/a-test-pkg/index.js
