+++ scripts/functions-discover-tests/fixtures/yarn-workspaces/packages/a-test-pkg/package.json
@@ -1,0 +1,5 @@
+{
+  "name": "@firebase/a-test-pkg",
+  "version": "0.0.1",
+  "private": true
+}
