--- scripts/functions-discover-tests/fixtures/yarn-workspaces/packages/a/package.json
@@ -1,4 +1,0 @@
-{
-  "name": "a",
-  "version": "0.0.1"
-}
