+++ scripts/functions-discover-tests/fixtures/yarn-workspaces/packages/functions/index.js
@@ -1,6 +1,6 @@
 const functions = require("firebase-functions");
 const { onRequest } = require("firebase-functions/v2/https");
-const { msg } = require("a");
+const { msg } = require("@firebase/a-test-pkg");
 
 exports.hellov1 = functions.https.onRequest((request, response) => {
   response.send(msg);
